Extract parser for Java version strings
In debug TestKit runs it is possible that the JavaVersion class is
loaded from the outdated TAPI distribution. The older JavaVersion may
lack newer version constants which makes checking for recent Java
versions unreliable.

This commit extracts version parser to use it without relying on
JavaVersion, thus avoiding the old TestKit pitfall.
